btrfs_supported_gemato: Add to tree
This adds a more advanced set of hooks, which are able to use btrfs to
manage the portage tree, working with gemato. It adds the following flow
when syncing:
	1) A snapshot of the portage tree is taken
	2) After syncing the authenticity of the portage tree is checked
	with gemato
	3) The following action depends on gematos result
		a) if checking the authenticity was successful, the
		snapshot is deleted and the updated portage tree is used
		b) if checking the authenticity failed the updated
		portage tree is rolled back to the previous snapshot

The machinery is adding a hook to sync_all_pre (for taking the
snapshot), sync_all_post (for running gemato and handling the snapshot).
Common code is added to python_helpers/btrfs.py. A configuration file,
which allows to specify for which repository a authenticity check should
be done, is expected to be located in
/etc/paludis/authenticated_repositories.json. By specifying
"method":"plain" in the json btrfs support is disabled and only gemato
checking is done for this repository.